0.3
Creation of all the methods and views to depending on the NewMessageSimpleView (EmailCreationView, SMSCreationView and TweetCreationView) and their ViewModels, to help the user to create a message more easily.
Creation of the ImportView and its ViewModel, to import messages from existing files (either .txt or .json) by choosing them of importing a complete folder content.

Testing: manual and auto